[modify] migrate 파일 생성 안되는 문제 해결, 타입 수

+++ migrations/versions/cc73bf0f602f_create_regions_table.py
@@ -20,7 +20,9 @@ def upgrade():
     op.create_table(
         "regions",
         sa.Column(
-            "id", sa.Integer().with_variant(sa.Integer(), "sqlite"), nullable=False
+            "id",
+            sa.SmallInteger().with_variant(sa.SmallInteger(), "sqlite"),
+            nullable=False,
         ),
         sa.Column("name", sa.String(length=50), nullable=False),
         sa.Column("created_at", sa.DateTime(), nullable=True),
